pyproject.toml
[project]
name = "unitree-launcher"
version = "0.2.0"
requires-python = ">=3.10,<3.11"
dependencies = [
    "numpy",
    "pyyaml",
    "onnxruntime>=1.16",
    "onnx>=1.14",
    "h5py",
]

[project.scripts]
sim = "unitree_launcher.main:cli_sim"
eval = "unitree_launcher.main:cli_eval"
real = "unitree_launcher.main:cli_real"
mirror = "unitree_launcher.main:cli_mirror"

[project.optional-dependencies]
sim = [
    "mujoco>=3.0",
    "viser>=0.2.0",
    "trimesh>=4.0",
    "matplotlib>=3.5",
    "cyclonedds==0.10.2",
    "unitree-sdk2py @ git+https://github.com/unitreerobotics/unitree_sdk2_python.git",
    "hidapi>=0.14",
    "mink>=1.0; python_version>='3.10'",
]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "pybind11>=2.11; sys_platform == 'linux'",
    "patchelf>=0.17; sys_platform == 'linux'",
]

[build-system]
requires = ["uv_build>=0.10.6,<0.11.0"]
build-backend = "uv_build"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
